requirements.txt
asgiref==3.3.1
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.1.5
django-cors-headers==3.6.0
django-filter==1.1.0
django-heroku==0.3.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-url-filter==0.3.15
djangorestframework==3.12.2
djangorestframework-filters==0.11.1
enum-compat==0.0.3
gunicorn==20.0.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.6
pytz==2020.5
requests==2.25.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
whitenoise==5.2.0
django-url-filter==0.3.15
dj-database-url==0.5.0
gunicorn==20.0.4
python-dotenv